service/requirements.txt
coqui-tts==0.26.2
torch
pyYAML
together
mlflow
python-box

